Add tests for items functions - part 2
