+++ test/integration/sugar/items/multi/upsert.test.js
@@ -53,6 +53,47 @@ describe("multi upsert items - client.table(table).multiUpsert()", function () {
     });
   });
 
+  describe("if the database is not empty", function () {
+
+    beforeEach(function () {
+      return client.create(testTable)
+    });
+
+    var items = [];
+    for (var i = 1; i < 188; i++) {
+      items.push({
+        id: "" + i,
+        email: i + "@epha.com"
+      });
+    }
+
+    var slice = items.slice(1, 27);
+
+    it("should properly insert all items and replace existing items", function () {
+      return client.table(testTable.TableName).multiUpsert(slice)
+        .then(function () {
+          items.forEach(function (item) {
+            item.checked = true;
+          });
+
+          return client.table(testTable.TableName).multiUpsert(items);
+        })
+        .then(function () {
+          return client.table(testTable.TableName).download();
+        })
+        .then(function (res) {
+          // test if arrays have the same length
+          expect(res).to.be.instanceof(Array);
+          expect(res).to.have.length(items.length);
+
+          // and now test if the items are equal
+          res.forEach(function (item) {
+            expect(items).to.contain(item);
+          });
+        });
+    });
+  });
+
   afterEach(function () {
     return client.remove(testTable)
       .catch(function () {
